feat: Simplify directions in dirReduc function
The dirReduc function simplifies a given array of directions by removing needless directions that cancel each other out. It uses a stack to keep track of the directions and checks if the current direction is opposite to the last direction in the stack. If they are opposite, the last direction is removed from the stack; otherwise, the current direction is added to the stack. The simplified directions are then returned.

Example usage:
const directions = ["NORTH", "SOUTH", "SOUTH", "EAST", "WEST", "NORTH", "WEST"];
dirReduc(directions); // returns ["WEST"]

Note: The code changes include the addition of the dirReduc function and the creation of the opposites object to map opposite directions.